refactor: Batch code optimization and simplification
1. FTP connection: Box `TlsStream` to fix FTPS connection; refactor `SYST` and `MDTM` command parsing logic
2. Utility methods: Replace `map_or` + `false` with `is_some_and` to simplify conditional checks
3. Iterators: Replace manual index loops with `iter().take()` and `zip()` to optimize iteration patterns
4. Calculation methods: Replace manual ceiling division with `div_ceil`, replace `max` + `min` with `clamp`
5. String handling: Replace `starts_with` + slicing with `strip_prefix` to simplify prefix checks
6. Array initialization: Replace manual loop initialization with `std::array::from_fn`
7. Code simplification: Merge duplicate branch conditions, remove redundant checks, simplify `Vec` initialization
8. Type optimization: Add `Default` derive to enums and remove manual implementations, simplify default value definitions
9. Formatting: Unify string concatenation patterns, remove unnecessary format wrappers
10. Documentation and refactoring: Fix comment errors, refactor DHT-related code types and logic

+++ aria2-core/src/dht/task.rs
@@ -491,7 +491,7 @@ impl DhtBucketRefreshTask {
         let buckets = routing_table.get_buckets();
         for bucket in &buckets {
             if self.force_refresh
-                || bucket.time_since_last_update().as_secs() > BUCKET_REFRESH_INTERVAL_SECS as u64
+                || bucket.time_since_last_update().as_secs() > BUCKET_REFRESH_INTERVAL_SECS
             {
                 // Generate a random ID within this bucket's range
                 let target = bucket.random_id_in_range();
